change examples folder structure
