--- examples/GeoSchelling/run.py
@@ -1,4 +1,0 @@
-from server import server
-import profile
-
-profile.run(server.launch())
