+++ examples/geo_schelling/data/nuts_rg_60M_2013_lvl_2.geojson
