+++ examples/geo_schelling/run.py
