+++ examples/geo_sir/README.md
@@ -15,8 +15,10 @@ Data obtained from [this link](http://adamw523.com/toronto-geojson/).
 
 ## How to run
 
-To run the model interactively, run ```mesa runserver``` in this directory. e.g.
-```
+To run the model interactively, run `mesa runserver` in this directory. e.g.
+
+```bash
 mesa runserver
 ```
-Then open your browser to [http://127.0.0.1:8521/](http://127.0.0.1:8521/), set the desired parameters, press Reset, and then Run.
+
+Then open your browser to [http://127.0.0.1:8521/](http://127.0.0.1:8521/) and press `Start`.
