+++ examples/geo_sir/data/TorontoNeighbourhoods-original.geojson
