+++ examples/geo_sir/data/TorontoNeighbourhoods.geojson
