+++ examples/geo_sir/run.py
@@ -1,0 +1,3 @@
+from server import server
+
+server.launch()
