+++ examples/geo_sir/server.py
@@ -1,8 +1,10 @@
 from mesa_geo.visualization.ModularVisualization import ModularServer
 from mesa.visualization.modules import ChartModule, TextElement
 from mesa.visualization.UserParam import UserSettableParameter
-from model import InfectedModel, PersonAgent
 from mesa_geo.visualization.modules import MapModule
+
+from model import GeoSir
+from agents import PersonAgent
 
 
 class InfectedText(TextElement):
@@ -62,4 +64,3 @@ server = ModularServer(
     "Basic agent-based SIR model",
     model_params,
 )
-server.launch()
